pyproject.toml
[project]
name = "agentic-trust-protocol"
version = "0.1.0"
description = "CQRS-based trust architecture for AI agents"
requires-python = ">=3.11"

dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "jupyter>=1.0",
]

[project.scripts]
naive-agent = "src.naive_agent.agent:main"
airlock-agent = "src.airlock_agent.agent:main"
run-eval = "eval.harness:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
